Add grouped Reddit adapter
